Increase max Hops
i was digging into the ubeswap codebase to see if i could programmatically get prices and i noticed that the useTrade hook has a MAX_HOPS const set to 3 

however the current farming incentives mean that the best trade (assuming farms with incentives are deepest) would for many be more than 3 hops. cETH --- MOO for instance is  CETH--MCUSD--CELO--MCELO--MOO  or cMCO2 to CBTC would be CMC02--UBE--CELO--MCUSD-CBTC  both being 4 hops.  or POOF to CUER POOF-UBE-CELO-MCEUR-cEUR